Add files via upload
Shouldn't forecast if correlogram partials and prior period correlations indicates the series may not be covariance stationary but model 2 was very close and the fitted on line tightly matched the realization so it was included anyway. Model 1 was fine but estimates appear a bit higher than I would expect. Consider the rates of change in gdp more telling in this repository!